Backport TCPIP instruction from SDL-Hyperion
This removes the facility support; for S370 the instruction is always
available in problem state: note that this is dangerous, as called out
in README.TCPIP. This fork of Spinhawk should only be used in controlled
environments where this risk is acceptable.

This builds on Unix-like systems; I do not know if the changes to
Makefile.am alone are enough to incorporate the new files into Windows
builds; I do not have a Windows development environment on which to test
and experiment.